Split into multiple files and clean up
